Use builder syntax and refactor tag registration
