+++ meta/src/link.rs
@@ -1,52 +1,8 @@
 use crate::use_head;
-use cfg_if::cfg_if;
 use leptos::*;
-use std::{
-    cell::{Cell, RefCell},
-    collections::HashMap,
-    rc::Rc,
-};
-
-/// Manages all of the Links set by [Link] components.
-#[derive(Clone, Default)]
-pub struct LinkContext {
-    #[allow(clippy::type_complexity)]
-    els: Rc<RefCell<HashMap<String, (HtmlElement<Link>, Scope, Option<web_sys::HtmlLinkElement>)>>>,
-    next_id: Rc<Cell<LinkId>>,
-}
-
-impl std::fmt::Debug for LinkContext {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("LinkContext").finish()
-    }
-}
-
-#[derive(Copy, Clone, Debug, Default, PartialEq, Eq, Hash)]
-struct LinkId(usize);
-
-impl LinkContext {
-    fn get_next_id(&self) -> LinkId {
-        let current_id = self.next_id.get();
-        let next_id = LinkId(current_id.0 + 1);
-        self.next_id.set(next_id);
-        next_id
-    }
-}
-
-#[cfg(feature = "ssr")]
-impl LinkContext {
-    /// Converts the set of Links into an HTML string that can be injected into the `<head>`.
-    pub fn as_string(&self) -> String {
-        self.els
-            .borrow()
-            .iter()
-            .map(|(_, (builder_el, cx, _))| builder_el.clone().into_view(*cx).render_to_string(*cx))
-            .collect()
-    }
-}
 
 /// Injects an [HTMLLinkElement](https://developer.mozilla.org/en-US/docs/Web/API/HTMLLinkElement) into the document
-/// head.
+/// head, accepting any of the valid attributes for that tag.
 /// ```
 /// use leptos::*;
 /// use leptos_meta::*;
@@ -126,8 +82,7 @@ pub fn Link(
     blocking: Option<String>,
 ) -> impl IntoView {
     let meta = use_head(cx);
-    let links = &meta.links;
-    let next_id = links.get_next_id();
+    let next_id = meta.tags.get_next_id();
     let id = id.unwrap_or_else(|| format!("leptos-link-{}", next_id.0));
 
     let builder_el = leptos::link(cx)
@@ -150,45 +105,5 @@ pub fn Link(
         .attr("type", type_)
         .attr("blocking", blocking);
 
-    cfg_if! {
-        if #[cfg(any(feature = "csr", feature = "hydrate"))] {
-            use leptos::document;
-
-            let element_to_hydrate = document()
-                .get_element_by_id(&id)
-                .map(|el| el.unchecked_into::<web_sys::HtmlLinkElement>());
-
-            let el = element_to_hydrate.unwrap_or_else({
-                let builder_el = builder_el.clone();
-                move || {
-                    let head = document().head().unwrap_throw();
-                    head
-                        .append_child(&builder_el)
-                        .unwrap_throw();
-
-                    (*builder_el).clone()
-                }
-            });
-
-            on_cleanup(cx, {
-                let el = el.clone();
-                let els = meta.links.els.clone();
-                let id = id.clone();
-                move || {
-                    let head = document().head().unwrap_throw();
-                    _ = head.remove_child(&el);
-                    els.borrow_mut().remove(&id);
-                }
-            });
-
-            meta.links
-                .els
-                .borrow_mut()
-                .insert(id, (builder_el, cx, Some(el)));
-
-        } else {
-            let meta = use_head(cx);
-            meta.links.els.borrow_mut().insert(id, (builder_el, cx, None));
-        }
-    }
+    meta.tags.register(cx, id, builder_el.into_any());
 }
